third_party/blink/web_tests/external/wpt/css/css-shapes/shape-functions/polygon-function-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Shapes Module Level 1: parsing the polygon() function</title>
<link rel="help" href="https://drafts.csswg.org/css-shapes-1/#funcdef-basic-shape-polygon">
<meta name="assert" content="Tests parsing of the polygon() function">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("shape-outside", "polygon(100px)");
test_invalid_value("shape-outside", "polygon(1%)");
test_invalid_value("shape-outside", "polygon(evenodd, 1px, 2px, 3em 4em, 5pt 6%)");

test_invalid_value("shape-outside", "polygon(round, 1px 2px)");
test_invalid_value("shape-outside", "polygon(round 1, 1px 2px)");
test_invalid_value("shape-outside", "polygon(round -1px, 1px 2px)");
test_invalid_value("shape-outside", "polygon(round 1%, 1px 2px)");
test_invalid_value("shape-outside", "polygon(round 45deg, 1px 2px)");
test_invalid_value("shape-outside", "polygon(round 1px 1px 2px)");
test_invalid_value("shape-outside", "polygon(1px 2px, round 1px)");
test_invalid_value("shape-outside", "polygon(round 1px round 2px, 1px 2px)");
test_invalid_value("shape-outside", "polygon(round 1px nonzero, 1px 2px)");
</script>
</body>
</html>
